added model.py
