+++ src/__init__.py
@@ -1,0 +1,6 @@
+from . import main
+from . import model
+from . import train
+from . import layers
+from . import attention
+from . import transformer
